Added the ability for entities to move about the world. Physics now has several methods for assisting in this movement.
Modified the 'HuntState' state of the FSM to cause the hunter to enter
into a wander.

TODO: Ensure that the hunter doesn't hunt too close or too far from his
clan center (this radius depends on the current state of the clan).

TODO: Rotate the entity in the direction of the velocity vector.