Primary hunting (really, gathering... might need to change the wording of the classes) is implemented:
Hunter Unit will search for resources by wandering through the world.
If a 'huntable' target is found, the unit will travel directly to it.
Once the unit arrives, he will gather a resource from the unit and
return it to his home (hardcoded (0, 0, 0).
As long as the last huntable target still has resources, the unit will
return, gather, then go home and deposit. Once the huntable target's
resources are depleted, the unit goes back to hunting.

TODO: The unit stops too abruptly. Needs to slow down to arrive at its
target realistically.

TODO: Implement a more complex inventory. Currently it's just an int.
This should be a new class.

TODO: Rendering of the inventory (boxes in front of the unit).

TODO: Rendering of the stockpile.

TODO: Tie the unit to its clan (the issue with this is some units don't
have clans... we can perform checks though).