The following TODOs from last time have been implemented:
TODO: Implement a more complex inventory. Currently it's just an int.
This should be a new class.

TODO: Rendering of the inventory (boxes in front of the unit).

TODO: Rendering of the stockpile.

TODO: Tie the unit to its clan (the issue with this is some units don't
have clans... we can perform checks though).